fix DI config
