+++ RDCore.SDK/Platform/IPlatformCompositionService.cs
@@ -2,12 +2,15 @@
 using Microsoft.Extensions.Logging;
 using Newtonsoft.Json.Linq;
 using OmniSharp.Extensions.LanguageServer.Protocol.Client.Capabilities;
+using RDCore.LanguageServer.Extensibility;
 using RDCore.SDK.Client;
 using RDCore.SDK.Extensibility;
 using RDCore.SDK.Server;
 using RDCore.SDK.Server.Services;
 using System.Collections.Immutable;
 using System.IO.Abstractions;
+using System.Reflection;
+using System.Text.Json;
 
 namespace RDCore.SDK.Platform;
 
@@ -17,6 +20,35 @@ public interface IPlatformCompositionService
     PlatformManifest GetManifest();
     ImmutableArray<ExtensionInfo> GetExtensions();
 }
+
+public class PlatformCompositionService(IFileSystem fileSystem, IExtensionsProvider extensions) : IPlatformCompositionService
+{
+    private static readonly string _manifestFileName = "rdcore.json";
+    private readonly string _rootPath = fileSystem.Directory.GetParent(Assembly.GetEntryAssembly()!.Location)!.FullName;
+    private PlatformManifest? _cached;
+    private ImmutableArray<ExtensionInfo>? _extensions;
+
+    public string RootPath => _rootPath;
+
+    public ImmutableArray<ExtensionInfo> GetExtensions()
+    {
+        _extensions ??= [.. extensions.Discover()];
+        return _extensions.Value;
+    }
+
+    public PlatformManifest GetManifest()
+    {
+        if (_cached is null)
+        {
+            var path = fileSystem.Path.Combine(_rootPath, _manifestFileName);
+            var content = fileSystem.File.ReadAllText(path);
+            _cached = JsonSerializer.Deserialize<PlatformManifest>(content)
+                ?? throw new InvalidOperationException();
+        }
+        return _cached;
+    }
+}
+
 
 /// <summary>
 /// An <em>abstract factory</em> that creates clients for platform components.
